:fire: 2024/09/27 초기화
