+++ src/main/java/io/hhplus/tdd/database/UserPointTable.java
@@ -1,12 +1,11 @@
 package io.hhplus.tdd.database;
 
+import io.hhplus.tdd.point.UserPoint;
+import org.springframework.stereotype.Component;
+
 import java.util.HashMap;
 import java.util.Map;
 import java.util.concurrent.TimeUnit;
-
-import org.springframework.stereotype.Component;
-
-import io.hhplus.tdd.point.dto.UserPoint;
 
 /**
  * 해당 Table 클래스는 변경하지 않고 공개된 API 만을 사용해 데이터를 제어합니다.
@@ -14,25 +13,25 @@ import io.hhplus.tdd.point.dto.UserPoint;
 @Component
 public class UserPointTable {
 
-	private final Map<Long, UserPoint> table = new HashMap<>();
+    private final Map<Long, UserPoint> table = new HashMap<>();
 
-	public UserPoint selectById(Long id) {
-		throttle(200);
-		return table.getOrDefault(id, UserPoint.empty(id));
-	}
+    public UserPoint selectById(Long id) {
+        throttle(200);
+        return table.getOrDefault(id, UserPoint.empty(id));
+    }
 
-	public UserPoint insertOrUpdate(long id, long amount) {
-		throttle(300);
-		UserPoint userPoint = new UserPoint(id, amount, System.currentTimeMillis());
-		table.put(id, userPoint);
-		return userPoint;
-	}
+    public UserPoint insertOrUpdate(long id, long amount) {
+        throttle(300);
+        UserPoint userPoint = new UserPoint(id, amount, System.currentTimeMillis());
+        table.put(id, userPoint);
+        return userPoint;
+    }
 
-	private void throttle(long millis) {
-		try {
-			TimeUnit.MILLISECONDS.sleep((long)(Math.random() * millis));
-		} catch (InterruptedException ignored) {
+    private void throttle(long millis) {
+        try {
+            TimeUnit.MILLISECONDS.sleep((long) (Math.random() * millis));
+        } catch (InterruptedException ignored) {
 
-		}
-	}
+        }
+    }
 }
