+++ src/main/java/io/hhplus/tdd/point/UserPoint.java
@@ -1,0 +1,12 @@
+package io.hhplus.tdd.point;
+
+public record UserPoint(
+        long id,
+        long point,
+        long updateMillis
+) {
+
+    public static UserPoint empty(long id) {
+        return new UserPoint(id, 0, System.currentTimeMillis());
+    }
+}
